pyproject.toml
[project]
name = "ocrpolish"
version = "0.1.0"
description = "A toolkit for cleaning, formatting, and validating OCR outputs processed by LLMs."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-docx",
    "ollama",
    "pydantic",
    "pyyaml",
    "click",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "N", "UP", "PL"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "--cov=ocrpolish --cov-report=term-missing"
